int -> double

+++ Fitness/src/main/java/by/mk_jd2_92_22/foodCounter/services/dto/ProductDTO.java
@@ -4,15 +4,15 @@ public class ProductDTO {
 
     private String name;
     private int kcal;
-    private int proteins;
-    private int fats;
-    private int carbohydrates;
+    private double proteins;
+    private double fats;
+    private double carbohydrates;
     private int weight;
 
     public ProductDTO() {
     }
 
-    public ProductDTO(String name, int kcal, int proteins, int fats, int carbohydrates, int weight) {
+    public ProductDTO(String name, int kcal, double proteins, double fats, double carbohydrates, int weight) {
         this.name = name;
         this.kcal = kcal;
         this.proteins = proteins;
@@ -37,27 +37,27 @@ public class ProductDTO {
         this.kcal = kcal;
     }
 
-    public int getProteins() {
+    public double getProteins() {
         return proteins;
     }
 
-    public void setProteins(int proteins) {
+    public void setProteins(double proteins) {
         this.proteins = proteins;
     }
 
-    public int getFats() {
+    public double getFats() {
         return fats;
     }
 
-    public void setFats(int fats) {
+    public void setFats(double fats) {
         this.fats = fats;
     }
 
-    public int getCarbohydrates() {
+    public double getCarbohydrates() {
         return carbohydrates;
     }
 
-    public void setCarbohydrates(int carbohydrates) {
+    public void setCarbohydrates(double carbohydrates) {
         this.carbohydrates = carbohydrates;
     }
